487: Applying refactoring to offers package
Applying the test refactoring to allow logic to be reused to call different API versions

Added createConsentAndGetAccessToken method to the AccountAccessConsent API as it is a commonly used pattern in the tests, updated existing tests to use this method where appropriate.

Issue: https://github.com/securebankingaccesstoolkit/securebankingaccesstoolkit/issues/487
