+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/accounts/api/v3_1_8/GetAccount.kt
@@ -4,10 +4,8 @@ import assertk.assertThat
 import assertk.assertions.isEqualTo
 import assertk.assertions.isNotEmpty
 import assertk.assertions.isNotNull
-import com.forgerock.securebanking.framework.configuration.psu
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
-import com.forgerock.uk.openbanking.support.account.AccountAS
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code
@@ -17,14 +15,7 @@ class GetAccount(version: OBVersion, tppResource: CreateTppCallback.TppResource)
     fun shouldGetAccountTest() {
         // Given
         val permissions = listOf(OBExternalPermissions1Code.READACCOUNTSDETAIL)
-        val consent = accountAccessConsentApi.createConsent(permissions)
-
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
+        val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
 
         val accountId = AccountRS().getFirstAccountId(accountsApiLinks.GetAccounts, accessToken)
