+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/balances/api/v3_1_8/GetBalances.kt
@@ -3,10 +3,8 @@ package com.forgerock.uk.openbanking.tests.functional.account.balances.api.v3_1_
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
 import assertk.assertions.isNotNull
-import com.forgerock.securebanking.framework.configuration.psu
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
-import com.forgerock.uk.openbanking.support.account.AccountAS
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code
@@ -19,13 +17,7 @@ class GetBalances(version: OBVersion, tppResource: CreateTppCallback.TppResource
             OBExternalPermissions1Code.READACCOUNTSDETAIL,
             OBExternalPermissions1Code.READBALANCES
         )
-        val consent = accountAccessConsentApi.createConsent(permissions)
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
+        val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
 
         // When
         val result = AccountRS().getAccountsData<OBReadBalance1>(
