+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/direct/debits/api/v3_1_8/GetDirectDebits.kt
@@ -3,12 +3,9 @@ package com.forgerock.uk.openbanking.tests.functional.account.direct.debits.api.
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
 import assertk.assertions.isNotNull
-import com.forgerock.securebanking.framework.configuration.psu
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
-import com.forgerock.uk.openbanking.support.account.AccountAS
 import com.forgerock.uk.openbanking.support.account.AccountRS
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code
 import uk.org.openbanking.datamodel.account.OBReadDirectDebit2
@@ -17,17 +14,11 @@ class GetDirectDebits(version: OBVersion, tppResource: CreateTppCallback.TppReso
     fun shouldGetDirectDebitsTest() {
         // Given
         val permissions = listOf(OBExternalPermissions1Code.READDIRECTDEBITS)
-        val consent = accountAccessConsentApi.createConsent(permissions)
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
+        val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
 
         // When
         val result = AccountRS().getAccountsData<OBReadDirectDebit2>(
-            accountAndTransaction3_1_8.Links.links.GetDirectDebits,
+            accountsApiLinks.GetDirectDebits,
             accessToken
         )
