+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/offers/api/v3_1_8/GetAccountOffers.kt
@@ -1,31 +1,35 @@
-package com.forgerock.uk.openbanking.tests.functional.account.beneficiaries.api.v3_1_8
+package com.forgerock.uk.openbanking.tests.functional.account.offers.api.v3_1_8
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
 import assertk.assertions.isNotNull
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
+import com.forgerock.uk.openbanking.support.account.AccountFactory
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
 import uk.org.openbanking.datamodel.account.*
 
-class GetBeneficiaries(version: OBVersion, tppResource: CreateTppCallback.TppResource): BaseAccountApi3_1_8(version, tppResource) {
-    fun shouldGetBeneficiariesTest() {
+class GetAccountOffers(version: OBVersion, tppResource: CreateTppCallback.TppResource): BaseAccountApi3_1_8(version, tppResource)  {
+    fun shouldGetAccountOffersTest() {
         // Given
         val permissions = listOf(
             OBExternalPermissions1Code.READACCOUNTSDETAIL,
-            OBExternalPermissions1Code.READBENEFICIARIESDETAIL
+            OBExternalPermissions1Code.READOFFERS
         )
         val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
+        val accountId = AccountRS().getFirstAccountId(accountsApiLinks.GetAccounts, accessToken)
 
         // When
-        val result = AccountRS().getAccountsData<OBReadBeneficiary5>(
-            accountsApiLinks.GetBeneficiaries,
-            accessToken
+        val result = AccountRS().getAccountsData<OBReadOffer1>(
+            AccountFactory.urlWithAccountId(
+                accountsApiLinks.GetAccountOffers,
+                accountId
+            ), accessToken
         )
 
         // Then
         assertThat(result).isNotNull()
-        assertThat(result.data.beneficiary).isNotEmpty()
+        assertThat(result.data.offer).isNotEmpty()
     }
 }
