+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/offers/api/v3_1_8/GetOffers.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.beneficiaries.api.v3_1_8
+package com.forgerock.uk.openbanking.tests.functional.account.offers.api.v3_1_8
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
@@ -9,23 +9,24 @@ import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
 import uk.org.openbanking.datamodel.account.*
 
-class GetBeneficiaries(version: OBVersion, tppResource: CreateTppCallback.TppResource): BaseAccountApi3_1_8(version, tppResource) {
-    fun shouldGetBeneficiariesTest() {
+class GetOffers(version: OBVersion, tppResource: CreateTppCallback.TppResource): BaseAccountApi3_1_8(version, tppResource)  {
+
+    fun shouldGetOffersTest() {
         // Given
         val permissions = listOf(
             OBExternalPermissions1Code.READACCOUNTSDETAIL,
-            OBExternalPermissions1Code.READBENEFICIARIESDETAIL
+            OBExternalPermissions1Code.READOFFERS
         )
         val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
 
         // When
-        val result = AccountRS().getAccountsData<OBReadBeneficiary5>(
-            accountsApiLinks.GetBeneficiaries,
+        val result = AccountRS().getAccountsData<OBReadOffer1>(
+            accountsApiLinks.GetOffers,
             accessToken
         )
 
         // Then
         assertThat(result).isNotNull()
-        assertThat(result.data.beneficiary).isNotEmpty()
+        assertThat(result.data.offer).isNotEmpty()
     }
 }
