+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/offers/legacy/LegacyGetOffersTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.offers
+package com.forgerock.uk.openbanking.tests.functional.account.offers.legacy
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
@@ -9,13 +9,12 @@ import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
 import com.forgerock.uk.openbanking.support.account.AccountAS
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.account.*
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READACCOUNTSDETAIL
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READOFFERS
 
-class GetOffersTest(val tppResource: CreateTppCallback.TppResource) {
+class LegacyGetOffersTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "accounts",
@@ -47,44 +46,6 @@ class GetOffersTest(val tppResource: CreateTppCallback.TppResource) {
         // When
         val result = AccountRS().getAccountsData<OBReadOffer1>(
             accountAndTransaction3_1_2.Links.links.GetOffers,
-            accessToken
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data.offer).isNotEmpty()
-    }
-
-    @EnabledIfVersion(
-        type = "accounts",
-        apiVersion = "v3.1.8",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetOffers"],
-        apis = ["offers"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5", "v.3.1.4", "v.3.1.3"]
-    )
-    @Test
-    fun shouldGetOffers_v3_1_8() {
-        // Given
-        val consentRequest = OBReadConsent1().data(
-            OBReadData1()
-                .permissions(listOf(READACCOUNTSDETAIL, READOFFERS))
-        )
-            .risk(OBRisk2())
-        val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_8.Links.links.CreateAccountAccessConsent,
-            consentRequest,
-            tppResource.tpp
-        )
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-
-        // When
-        val result = AccountRS().getAccountsData<OBReadOffer1>(
-            accountAndTransaction3_1_8.Links.links.GetOffers,
             accessToken
         )
